requirements.txt
anyio==4.9.0
asgiref==3.8.1
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
distlib==0.3.9
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pipenv==2025.0.2
platformdirs==4.3.8
pymongo==4.13.0
python-dotenv==1.1.0
setuptools==80.9.0
sniffio==1.3.1
virtualenv==20.31.2
waitress==3.0.2
Werkzeug==3.1.3
